requirements.txt
requests
pillow==5.0.0
webcolors
